--- src/main/java/roomescape/entity/Role.java
@@ -1,6 +1,0 @@
-package roomescape.entity;
-
-public enum Role {
-    USER,
-    ADMIN
-}
