--- src/main/java/roomescape/exception/ForbiddenException.java
@@ -1,8 +1,0 @@
-package roomescape.exception;
-
-public class ForbiddenException extends RuntimeException {
-
-    public ForbiddenException(String message) {
-        super(message);
-    }
-}
