--- src/main/java/roomescape/exception/InvalidMemberException.java
@@ -1,8 +1,0 @@
-package roomescape.exception;
-
-public class InvalidMemberException extends IllegalArgumentException {
-
-    public InvalidMemberException(String message) {
-        super(message);
-    }
-}
