--- src/main/java/roomescape/exception/InvalidReservationException.java
@@ -1,8 +1,0 @@
-package roomescape.exception;
-
-public class InvalidReservationException extends IllegalArgumentException {
-
-    public InvalidReservationException(String message) {
-        super(message);
-    }
-}
