--- src/main/java/roomescape/exception/InvalidReservationTimeException.java
@@ -1,8 +1,0 @@
-package roomescape.exception;
-
-public class InvalidReservationTimeException extends IllegalArgumentException {
-
-    public InvalidReservationTimeException(String message) {
-        super(message);
-    }
-}
