--- src/main/java/roomescape/exception/InvalidThemeException.java
@@ -1,8 +1,0 @@
-package roomescape.exception;
-
-public class InvalidThemeException extends IllegalArgumentException {
-    
-    public InvalidThemeException(String message) {
-        super(message);
-    }
-}
