--- src/main/java/roomescape/exception/UnauthorizedException.java
@@ -1,7 +1,0 @@
-package roomescape.exception;
-
-public class UnauthorizedException extends RuntimeException {
-    public UnauthorizedException(String message) {
-        super(message);
-    }
-}
