--- src/main/java/roomescape/global/dto/ErrorResponse.java
@@ -1,10 +1,0 @@
-package roomescape.global.dto;
-
-public record ErrorResponse(
-        String message
-) {
-
-    public static ErrorResponse from(String message) {
-        return new ErrorResponse(message);
-    }
-}
